# encoding: Avoid initially setting decimal value to 0 when decoding