This squeezes a little extra performance out of
`DecodeDecimalAscending`.

```
name             old time/op    new time/op    delta
DecodeDecimal-4     238ns ± 1%     220ns ± 1%  -7.64%          (p=0.000 n=9+9)

name             old alloc/op   new alloc/op   delta
DecodeDecimal-4     96.0B ± 0%     96.0B ± 0%    ~     (all samples are equal)

name             old allocs/op  new allocs/op  delta
DecodeDecimal-4      2.00 ± 0%      2.00 ± 0%    ~     (all samples are equal)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/5392)
<!-- Reviewable:end -->